refactor(engine) #5686: keep CypherReferencedVariables beside the AST it analyses
Review point: CypherStatement named a parser type to state its own contract, which is a package edge the ast
package did not have before. The class analyses these nodes, is returned by an ast interface and memoized on an
ast implementation, so ast is where it belongs; moving it takes the parser type off the contract entirely.

What it still reaches for is CypherExpressionWalker, a traversal over these same nodes filed under the parser
that first needed it. Said so at the class rather than left to be rediscovered.

Also records two limits the review named: a UnionStatement body permanently answers "unknown" and never recovers
a push-down (safe, and no UNION can satisfy either push-down anyway), and the reflection guard reads declared
types, so an expression stashed in a raw Object field would look like a leaf.
